"6:20pm. I am back. Today I would have gotten all of this done had it not been for the constant weather related interruptions.
6:25pm.

```fs
| BundlePrototype(_,(_,name),vars,expr) -> // TODO: Validate that there are no unused foralls.
```

I'll do this thing tomorrow. Right now after all the time I had to think about how to do the instance, the best way to do things is extremely clear to me. Tomorrow, I'll do the forall validation and the instances and then I'll take some time to the the review.

Alternatively, I'll leave the review for the day after.

6:30pm. Actually I am pretty tired right now. The first part of the day was brilliant and I really handled constraint processing like a pro. It really shows how important it is to put in effort every day. You only have so many good hours to put into programming out of however many you are awake.

6:35pm. Tomorrow is a big day. After I finish this, the typechecker will be over 95% complete. If that is not a good way to end the month, I do not know what is."