"3:40pm. https://www.reddit.com/r/reinforcementlearning/comments/ihu34l/impact_of_go_ai_on_the_professional_go_world/g35qxb8?utm_source=share&utm_medium=web2x&context=3
The Magnus Carlsen posts are pretty interesting.

4:15pm. Enough wasting time. Let me do the review for a bit.

///

At the time of writing, I've just finished putting in prototypes (typeclasses) and the typechecking stage is 95% complete. It is integrated with editor support, and hover works. What is left to do is testing, and some minor features that will be easy to put in. Things are going really great. My ETA until [Spiral v0.2](https://github.com/mrakgr/The-Spiral-Language/tree/v0.2) is in a usable state is two months. The work I've done in the past month was novel to me - I've never implemented top-down type checking before. All the code had to be done from scratch, while the passes that are left I have deep familiarity with, so they won't give me much trouble.

Since dreams get adjusted in the face of reality, I had to leave out various features I had originally envisioned such as multi argument constraints and higher ranked types. GADTs are likewise out as they would play poorly with the later phases. GADTs in particular would be easy to implement in the partial evaluator, but despite that even if I had decided to go through significant difficulties of implementing them in the top-down segment, the kind of code resulting from that would not be a straightforward match to the bottom-up segment. So they are out - what Spiral has now is what it will have from here on out.

I had to brainstorm quite a lot on the design. For example, I really wanted named tensor dimensions and spent a long time thinking how to fit them into the top-down segment. For a long time I've been convinced that I would need multi argument constraints, but eventually I figured out how to do them without which was a significant relief. It would have been really difficult to do multi arg constraints within the confines of the current design.

The two libraries I am using to do editor support, NetMQ and Hopac are really proving their worth. Without the concurrency primitives afforded by them the task would be much harder. I spent a long time studying them and that effort is paying off now. More than just editor support for a language, having an understanding of them significantly increases the range of useful programs I can write and the exercise in concurrency will no doubt be good preparation for when I get my hands on the [neurochips](https://cacm.acm.org/magazines/2020/8/246356-neuromorphic-chips-take-shape/fulltext) I am coveting. [Future hardware](https://www.080910t.com/downloads/AI%20Chip%202018%20 EN.pdf) is what I am casting my net for. Thanks to the novel programming staged functional style enabled by Spiral, I am unusually well suited for taking advantage of all such developments.

I am looking forward to it. Being able to use Spiral like a real language will close the productivity gap that chased me away from v0.09 and I will be able to use my real power regardless of the domain.

Some people think that the hardware that is going to come will be a huge change in terms of programming requirements, but I am not convinced. Whatever distributed programming challenges there are in using neurochips, it is very unlikely than any of that will be harder than communicating over the Internet. Concurrency is not a new thing in the programming world.

What is really important that I operate at my peak capacity. No matter what, I do not want to give up that feeling. In terms of (top-down) type system features, Spiral does not stand out at all - though it a very capable functional programming language. It cannot compete with Haskell or Scala when it comes to type systems. Rather, what it allows functional programmers is for them to stick to their principles. The best Assembly programmer in the world would at most be as good as a mediocre C programmer. The relationship is similar between best C programmers and mediocre functional programmers.

Since I want to be good overall, it is important can exhibit what makes me good when writing things like GPU kernels. It is important that when doing a ML library, I have seamless integration between all the pieces of hardware. It is important to me to be able to do functional programming even in domains where GC is not an option. It is important to me that the C segments do not become a bottleneck and that I do not waste my time writing wrappers. It is important to me that I bring out the full composability of the underlying design. I know for a fact that this cannot be done in F# or Haskell or Scala.

It has to be a design like Spiral's. The user of the language has to have a notion of what the partial evaluator is doing under the hood. The language has to have join points as they are the best way of getting a grip on such a system, they are what makes partial evaluation composable. They are the bottom-up equivalent of top-down type annotations.

A language which allows reliable and accurate control over inlining is the one that can smoothly handle the trade off between heap allocation and code size.

Right now, it is too late for me to revolutionize GPU computing as it is too mature. Had Spiral v0.2 been mature in 2010 it might have made significant impact in that domain, but I am going to show just how big of an advantage not needing to drop to the C level really is on this next batch of hardware. There is actually a huge cost in both time and effort needed to make high quality ML libraries, and I am going to cut those costs down to but a fraction. After that I am going to do what I tried in 2018, except much better.

///

5:35pm. I think this will do. Let me go with this. When the thread come up in a day or two I'll post this.

Now I am done for the day.

Tomorrow, what is waiting for me is testing. I will finally get the chance to try all of this for real. Honestly, I am so happy now it almost feels like a reward. I'll finally be able to program in Spiral v0.2. This is the language with which I will eat the universe. Conquering the world is not enough.

5:40pm. Tomorrow will be a small echo of the future when will be doing the kind of programming I wanted to for real.

2018 was just not enough. It was not enough to satisfy me.

In 2021 I will get my chance again. This time I will play it extemely cheap. Rather than tossing away known advantages in order to build my skills, I will just cheese my way through using ensembles. There is no need to think too hard about credit assignment. I'll play it like a god and just weight the probabilities accordingly.

5:45pm. Damn it, I really want it. I am finally getting into it for real. I wrote this all this type inference code, but hadn't had the chance to use it yet. I want a bigger bite of that. Right now the memory of putting a cursor over a number and seeing it say `i32` is just the sweetest thing in the world.

Just how much did I have to suffer and learn to be able to get this simple thing. These new skills I've acquired in 2020, you can bet I will be taking advantage of them in spades.

5:50pm. Ahhh, the Singularity...I keep searching for it, but maybe the truth is that I am in the middle of it. I mean, rare amongst humans I can see that all the good ways of getting real life superpowers involve identity grinding bullshit. You either integrate with the machines with sheer force of will and detemination, or you stand on the sidelines with the rest of the non-programming characters. They have almost complete overlap with NPCs.

Whatever misgivings I might have and whatever mistakes I have made, the moto of 'anything for omnipotence' directs my effort towards the black maw that is the future."