"7:30pm. Change of plans.
```fs
        let rec associate_prot_foralls_with_ins_names ty = function
            | (a : Var) :: a', (((r,(n,_)),_) : TypeVar) :: b' ->
                let x = TyVar a
                hover_types.Add(r,x)
                associate_prot_foralls_with_ins_names (Map.add n x ty) (a', b')
            | [], ((r,_),_) :: b' -> errors.Add(r,ExcessForallInInstance); associate_prot_foralls_with_ins_names ty ([], b')
            | _, [] -> ty
```

I made a mistake with the way I handle foralls.

Instead of doing this thing, what I should do is make the foralls completely implicit. I should look at the prototype and just add those to the environment directly.

```fs
            assert_no_kind ins_foralls
            assert_no_constraint ins_foralls
```

I am already asserting that it should not have kind and constraint annotations. Why not get rid of these entirely.

Why is this so troublesome?

When I start filling in the missing type application I will rely on the `name` field. Having a forall whose binding does not match is name like in `associate_prot_foralls_with_ins_names` will break that.

But I also do not want to have to assert that foralls should have equal names.

It did not occur to me earlier in the day, but the solution is to just get rid of names entirely.

```fs
            let prototype = top_env'.prototype.[i]
            hover_types.Add(fst prot, prototype.signature)
```

Let me do this.

7:45pm.

```fs
    | BundleInstance(r,prot,ins,vars,body) ->
    | BundleInstance(r,prot,ins,vars,body) ->
        let body i (i',k) =
            let er_count = errors.Count
            let guard next = if errors.Count = er_count then next () else top_env'
            let ins_kind = kind_get k
            let prototype = top_env'.prototype.[i]
            hover_types.Add(fst prot, prototype.signature) // TODO: Do the same for the instance signature.
            let prot_kind = kind_get prototype.kind_domain
            if ins_kind.arity < prot_kind.arity then errors.Add(r,InstanceArityError(prot_kind.arity,ins_kind.arity))
            guard <| fun () ->
            let vars_expected = ins_kind.arity - prot_kind.arity
            let _ =
                let got = List.skip vars_expected ins_kind.args |> List.reduceBack (fun a b -> KindFun(a,b))
                let expected = prototype.kind_domain
                try unify_kind' (fun () -> raise (TypeErrorException [r, InstanceKindError (got, expected)])) got expected
                with :? TypeErrorException as e -> errors.AddRange e.Data0
            guard <| fun () ->
            let vars_count = List.length vars
            if vars_count <> vars_expected then errors.Add(r,InstanceCoreVarsShouldMatchTheArityDifference(vars_count,vars_expected))
            guard <| fun () ->
            vars |> List.iter (fun ((r,(_,k)),_) -> match k with RawKindWildcard -> () | _ -> errors.Add(r,KindNotAllowedInInstanceForall))
            match body with RawForall(r,_,_) -> errors.Add(r,ForallsNotAllowedInInstanceBody) | _ -> ()
            guard <| fun () ->
            let ins_core, env_ty =
                List.mapFold (fun s (((r,_),_) & x,k) ->
                    let v = {typevar_to_var s x with kind = k}
                    let x = TyVar v
                    hover_types.Add(r,x)
                    x, Map.add v.name x s
                    ) Map.empty (List.zip vars ins_kind.args)
            let trim_kind = function KindFun(_,k) -> k | _ -> failwith "impossible"
            let ins_core_ty,_ = List.fold (fun (a,k) (b : T) -> let k = trim_kind k in TyApply(a,b,k),k) (TyHigherOrder(i',k),k) ins_core
            let env_ty, prot_body =
                match foralls_ty_get prototype.signature with
                | (prot_core :: prot_foralls), prot_body ->
                    List.fold (fun ty x -> Map.add x.name (TyVar x) ty) env_ty prot_foralls,
                    subst [prot_core, ins_core_ty] prot_body
                | _ -> failwith "impossible"
            term {term=Map.empty; ty=env_ty} prot_body body
            let prototype = {|prototype with instance = Map.add i' (ins_core |> List.map (function TyVar x -> x.constraints | _ -> failwith "impossible")) prototype.instance|}
            {top_env' with prototype = PersistentVector.update i prototype top_env'.prototype}
```

This is actually a lot better. This is excellent design all around.

```fs
| ForallsNotAllowedInInstanceBody -> "Foralls are not allowed in instance bodies. The variables are already bound implicitly and are present in the environment."
```

Brilliant. Because of this, I do not have to worry about refactoring suddenly breaking anything. The user will be able to rename...agh, actually it is the same thing.

Well, whatever. If the instance vars break, then let them. The only restriction I am going to have to add is for the explicit vars not to intersect with the prototype ones.

```fs
                    List.fold (fun ty x ->
                        vars |> List.iter (fun ((r,(a,_)),_) -> if a = x.name then errors.Add(r,InstanceVarShouldNotMatchAnyOfPrototypes))
                        Map.add x.name (TyVar x) ty) env_ty prot_foralls,
```

Let me do this.

```fs
| InstanceVarShouldNotMatchAnyOfPrototypes -> "Instance forall must not have the same name as any of the prototype foralls."
```

8pm. I know - this will make renaming prototype foralls a nightmare, but it is not like what I had before would have worked either. It would have worked purely for typechecking purposes, but I would have had real trouble in the later passes.

This kind of arrangement should be strong incentive not to juggle forall vars explicitly in code. Let the TC automatically fill them in.

Still, the worst that can come from this is that if the user poorly names the forall vars, at worst they will be `a`, `b` or `x`. It is not the end of the world.

8:05pm. Let me close here again.

I am not inspired right now. If there are any mistakes, I'll catch them during testing tomorrow.

The current design's only benefit might be that it is easier to implement, but that is enough benefit to make this worth it to me."