"2:30pm. Done with breakfast. Let me read Demon Maiden for a bit and then I will start.
2:40pm. https://boards.4channel.org/g/thread/77651414/soup-g

///

Soup /g/

So I went for a tech interview today, and was doing great until I got this question that stumped me; I panicked and fucked up.

Write a solution that will return the numbers of ways you can sum to n using no more than any two numbers in the form of 2^n (i.e. for n=10, you can use at most 2x1, 2x2, 2x4, 2x8)

Order doesn't count.

Examples:

5 -> 2 ([4+1], [2+2+1])
10 -> 5 ([8+2], [8+1+1], [4+4+2], [4+2+2+1+1], [4+4+1+1])
47 -> 2 ([32+8+4+2+1], [16+16+8+4+2+1])

So boom, I wrote a brute force and was proud of myself. And then the interviewer asked me to enter a large number.

Cringe as my shit solution span and span and span, and the interviewer said there's a solution using caching that will make the computation tiny.

They gave me a take-home to try and finish it BUT I'm convinced this shit is fucked.

Should I just ficking quit, or is interviewer full of shit? Feels bad man. CS degree was worthless :(

///

Now, I am thinking about this thing.

2:45pm. No, I am not going to do HackerRank problems here.

...I really do not understand the question as posed.

///

5 -> 2 ([4+1], [2+2+1])

///

Shouldn't [2+1+1+1] be a possiblity here?

Ah, you can't use more than two of any number.

2:50pm. Ok, I get it.

...And I know how to code up an efficient solution for this.

2:55pm. Ok, I see it most of the way. For a minute there I would be stumped at this problem, but if popped up I could do it.

Let me move on.

Let me finish that chapter.

3:40pm. Ok, just a bit more and then I am going to get into it.

4pm. Ok, I am satisfied with the degree of wastage during the break. I can't believe it lasted this long.

Let me at least crack open the prepass.

```
//let pattern_to_rawexpr (arg: VarString, clauses: (Pattern * RawExpr) []) =
```

I have this thing waiting for me there, commented out.

4:05pm. Ok, first of all...

Originally I intended to use `RawExpr`, the same thing I've parsed and typechecker when filling in the type applications. But I've realized that this has an issue.

```
inl q forall t. (x : t) =
    inl w forall t. (y : t) : bool = $"(=)" x y
    w x
```

Suppose I have a macro like this.

Obviously I need to fill in the type annotation for it.

What should the type of `$"(=)"` be? `t -> t -> bool`?

That is the crux of the issue. I can't put `t` there without that causing a name collision.

The way to get around that is to do the filling while at the same time replacing the names of all the variables with their id.

4:15pm. I am going to have to come up with new types for everything. Not just the thing after `RawExpr`. I am going to have to replace `T` with something whose vars have ids.

5:15pm. I am really thinking at full blast about this. For the first run, I am going to get rid of one of the optimizations. I am going to have the let statements have their tags right there. The same goes for tests. It is just too much of a pain in the ass to do anything else in the prepass environment.

5:25pm. Did I put macros in type application?

```fs
        | RawMacro(_,a) ->
            type_application.Add(x,s)
```

As it turns out, I forgot this.

5:30pm. Also I need to do a check so that raw macros never show up alone.

```fs
| RawMacro(r,a) -> errors.Add(r,MacroIsMissingAnnotation); cmacro term ty a
```

This in the check function. I'll next put some special behavior in `RawAnnot`.

5:35pm.

```fs
        | RawAnnot(_,RawMacro(_,a),b) -> cmacro term ty a; ctype term ty b
        | RawMacro(r,a) -> errors.Add(r,MacroIsMissingAnnotation); cmacro term ty a
        | RawAnnot(_,a,b) -> cterm term ty a; ctype term ty b
```

Let me do it like so.

5:50pm. Ah, I see. Rather than having two different arrays or just checking immediatelly, I can distinguish recursive from non-recursive variables by having the recursive ones have negative ids. This would just be a small optimization.

This is the ideal choice.

6pm. https://www.youtube.com/watch?v=Jpu7jHRRIn4
Mick Gordon - Mars Core (DOOM Eternal - Extended Gamerip)

Thinking about programming while listening to this. It is my favorite track, but it plays for only a single battle.

I think I will call it a day here. Today was slow.

Redesigning the servers took most of my energy today, but I did spend whatever was left wisely.

6:15pm. I am still thinking about it.

The rule that or patterns need to have the same variables on both sides will make dealing with them easier.

I am changing my mind on the whole RawInline thing. I am going to do something special on order to memoize the different pattern branches, but otherwise I am going to remove them during the second pass.

Since I am making a fresh AST, I might as well merge the inline with joint points.

6:25pm. Right now I only have one kind of join point, but I think it might be worth introducing `join'` so I can write something like `join' cuda`. It would take an argument specifying its type. Doing it like this is good. I really should not have to wrap all the join points in `RawInline`.

6:35pm. How about I really stop for the day here. I am still obsessed about it and that is good, but after a while I need to force myself to stop.

6:40pm. I am going through the next part over and over in my head, and I think that I can deal with it as expected. Now that I am touching upon the pieces and turning them over, I see that they all fit together quite nicely.

I can do this.

Not only am I confident. Before I started thinking about this I got the feeling that the prepass might take me 1-2 weeks, but now it feels like I can deal with it much faster than that. I should be able to mostly finish it in the next few days.

Of course now that I need to integrate it with the server, things will go more slowly, but still, this is not the kind of obstacle that I will be stuck on for too long.

6:45pm. I am motivated. I am going to crush this thing tomorrow. Getting the prepass out of the way will be a huge relief as after that only the partial evaluator will remain. In terms of difficult, codegen barely registers in my mind."