"2:25pm. Took a little break. The reason why I need those foralls and inl is because...Actually, I do not need the forall. Not at the type level.
```fs
and TExpr =
    | B of Range
    | Var of Range * Id
    | Pair of Range * TExpr * TExpr
    | Arrow of Range * TExpr * TExpr
    | Record of Range * Map<string,TExpr>
    | Symbol of Range * string
    | Apply of Range * TExpr * TExpr
    | Prim of Range * Config.PrimitiveType
    | Term of Range * Expr
    | Macro of Range * Macro list
    | Fun of Range * Id * TExpr
```

Type level on the other hand I will need.

For type level functions, should I just do explicit substitution?

No, it has to be implicit.

```fs
| Fun of Range * FreeVars * Id * TExpr
```

Then let me take care of them properly.

2:40pm.

```fs
type Macro =
    | MacroText of Range * string
    | MacroTypeVar of Range * Id
    | MacroTermVar of Range * Id
type RecordWith =
    | RecordWithSymbol of (Range * string) * Expr
    | RecordWithSymbolModify of (Range * string) * Expr
    | RecordWithInjectVar of (Range * Id) * Expr
    | RecordWithInjectVarModify of (Range * Id) * Expr
and RecordWithout =
    | RecordWithoutSymbol of Range * string
    | RecordWithoutInjectVar of Range * Id
and PatRecordMember =
    | PatRecordMembersSymbol of (Range * string) * Id
    | PatRecordMembersInjectVar of (Range * Id) * Id
```

I don't need some of the ranges here.

...Well, nevermind. I'll leave them in for now. Let me not remove them until I really need to do that.

2:50pm.

```fs
| PatternMemo of Range * Expr // Acts like a join point during the prepass. Should not appear during peval.
```

I need the inline here after all. This won't cut it.

I though that since I have the variable disjointedness rule that I might need need different paths thought the pattern matcher, but unfortunately, during peval time the tags might be different. So I have to account for that during the prepass.

```fs
| Inline of Range * FreeVars * Expr // Acts like a join point during the prepass.
```

Let me go with this. This is the usual form.

Agh, I am going to have to add `RawInline` as well. This won't work like this.

3pm. Adding it wasn't too hard. Ok.

3:05pm. It seems like I am not doing anything here, but I am actually in a state of high tension. How about I step away from the screen for a while? Yeah, it is back to this behavior again.

I need to embrace my loneliness a little. I need to find the desire to do what needs to be done. If I can do the prepass properly, the partial evaluator will be much easier. Thanks to all the extra work that I need to do this time around, it won't be as easy as it was last time.

```fs
| TypeHigherOrderDestruct of Range * arity: int * bind: Id * on_succ: Expr * on_fail: Expr
```

Rather than doing the arity here, I should base it the number of binds.

```fs
| TypeHigherOrderDestruct of Range * bind: Id * pat: Id list * on_succ: Expr * on_fail: Expr
```

Yeah, this is the way to do it.

3:20pm.

```fs
| TypeHigherOrderTest of Range * ho: Id * bind: Id * on_succ: Expr * on_fail: Expr
```

Forgot it is no longer just nominal.

3:30pm. Let me take that break for the last of the day. I will just go to bed and spent my time that way. I'll take the chance to have a bath while I am at it.

I need some time to prime myself. Once I got through the rough spots, the typechecker went by without a hitch. I need to go through the same motions for the prepass."