"6:20pm. I'll stop here for the day. Tomorrow, I might just spend the entire day in bed brainstorming. Right now, I am not even thinking that much what I need to do. Rather I am just trying to give myself a kick.
I need to enter the mental state where doing the prepass is the only thing in my life. Forget about wanting to do it quickly. It is fine if it takes a few weeks. Whatever is needed is fine. Do it slowly until you are satisfied. What is important that I work on it. What is important the the prepass is important to me.

Once something is trully important in a person's life, that something will be driven to completion.

6:35pm. But you also do not want thing important to you to be a burden.

So it is best to do things slowly and let the momentum build on its own. Yeah, I'd much rather program in Spiral than work on its compiler, but I have to get the bad out of the way before I can get the good. It is a kind of training. Athletes train before they compete, but unlike athletes who have to maintain their body all year round the things I carve out into code will stay that way.

Spiral will last me a long time.

Past Spiral v0.2, the language innovations go up a level in difficulty where any single person would be tied up for years working on them. The only way trying to grasp such high hanging fruit makes sense if you are Skynet and can do in a minute what would take a human in a year - or you are also Skynet and have infinite lifespan.

So I am not really worried about falling behind. There are features like effects, depends types and such, but by the time humans are ready to exceed Spiral v0.2 I will have long have gotten my power.

6:40pm. Working in top down Spiral will be my reward.

I reallly need to get my hands on those neuromorphic chips. GPUs are just not enough. They are good practice for making a language, but they were not a realistic option from the start.

6:45pm. Since it is training, I see it as my goal to attack from a weak position and increase the difficulty artificially, but I am going to break that habit with v0.2.

6:50pm. As I look back on my life, I get such a great feeling of fatigue. I've had enough training. I want the fun stuff.

For that I need to complete the prepass first.

Tomorrow, I am seriously going to try and get something going. Though the passes will be fused, I should just assume that pattern and typecase compilation is given and the rest. I can also do free variable gathering with no problem. Trying to do it all at once it too tedious. So I should just concentrate on completing a subgoal. Once I do that, I'll be able to gather some steam. It does not matter even if I can't do the subgoal or if I can't gather the motivation.

If I can't do the goal itself, then that is fine. I'll spread it over several days. If I can't gather the motivation despite having a well defined goal, I'll just go to bed and use the usual motivation gathering technique.

7pm. It is really impossible to get started on something when it feels like a huge mountain. Many people want to climb big mountains and then they get tired very quickly at the outset. Great men dream of reaching the pinnacle, but they understand that the next step is what is really important."