"8:35pm. Apart from having lunch, I was in bed for this long.
8:40pm. I'll admit the first 2/3rd was not very fruitful and my insecurities were just eating away at me, but in the past hour, I feel like I am making very real breakthroughs.

All this seems easy, I mean I know how to do renaming, but now that I have to do filling alongside it, the pieces are harder to fit. I know how to do this now. I even know how to get around the need for RawInline in patterns. I was really confused about this earlier, but here is what I can do.

```
inl f a (A: _,_b | B: b,_) = ...
```

Earlier I realized that I could not just tag a as 0 and b as 1 because those wildcards and intermediates would get in the way. In the previous version of the prepass I used `RawInline` to get around this problem, but what I am going to do now is count the wildcards separately. Those dummy vars the pattern match compiler introduces will have negative scope starting at -1. Something that is really good about all of this is that the wildcards are very contained, so if I move from the scheme where I am just pushing vars forward, I will be able to reuse that space on the stack.

In fact, in order to be able to tag `a` as 0 and `b` as 1 I will have to use the scheme where let statements and tests take in the binding number.

8:55am. Ok, that are essentially the properties of the first pass. I'll use 3 passes.

1) The first one - filling and gathering of free vars and pattern compilation.
2) The second one is for resolving recursive variables. I am not going to be able to do the scheme using global dictionaries that I had in mind. Instead I'll do a pass that does it top to bottom. Maybe that is for the best.
3) The downwards renaming for efficiency that the partial evaluator needs. I'll also calculate the stack size during this point by getting the max var.

9:05pm. Actually, I forgot that I need to strip pattern annotations during the first pass as well. And can't forget to fill in the functions.

Yeah, it is going to be a busy one.

There is a shitload of cases to cover, so much more than before. And actually I need to inline the globals as well, but need to leave the stack size out during the first two passes.

I am going to have to replace `FreeVars` with a generic in order to handle this. Then the global will be a specific instance of such an AST that I'll be able to include.

9:15am. The amount of work to be done is fairly large. The prepass is no longer a trivial thing by any means. The steps are numerous and it will take a while.

Let me stop here. I'll continue brainstorming tomorrow."