"9:30am. Yesterday I ended being in bed since the 4:20pm.
I've figured out where a major part of my demoralization comes from.

```fs
let prepass (top_env : TopEnv) expr =
    let v_term (env : Env) x = Map.tryFind x env.term.env |> Option.defaultWith (fun () -> top_env.term.[x])
    let v_ty (env : Env) x =  Map.tryFind x env.ty.env |> Option.defaultWith (fun () -> top_env.ty.[x])
```

It is this stupid thing. It is so heavy. The way it is written now is completely trivial though. So what is the problem? Why did making this feel like I've been wading through a swamp?

The issue is the filling.

Let me quote myself.

> Once I have that, I'll be able to move on to compliing patterns and typecase. Then comes free vars propagation. Then comes filling. Then comes recursive var resultion. Then comes name lowering.

The filling is somewhere in the middle mentally. And that is a problem. What do I need to do the filling?

This dictionary from `infer`.

```fs
let type_application = Dictionary(HashIdentity.Reference)
```

I define it right near the start.

```fs
    | BundleInstance(r,(_,prot),(_,ins),l,body) ->
        let prot_id = failwith "TODO"
```

I will also need a way of passing the prototype ids from the top level. Meaning I need to have a dependency on the inferencer here as well.

All this is weight. Huge weight.

It is not like going from parsing to type inference. The type inference part is essentially incomplete by the time it gets to the prepass.

It is not like it is in some ready made form either.

```fs
and T =
    | TyB
    | TyPrim of PrimitiveType
    | TySymbol of string
    | TyPair of T * T
    | TyRecord of Map<string, T>
    | TyFun of T * T
    | TyArray of T
    | TyNominal of int
    | TyUnion of Map<string,T>
    | TyApply of T * T * TT // Regular type functions (TyInl) get reduced, while this represents the staged reduction of nominals.
    | TyInl of Var * T
    | TyForall of Var * T
    | TyMetavar of MVar * T option ref
    | TyVar of Var
    | TyMacro of TM list
    | TyLayout of T * Layout
```

I haven't decided how I am going to convert this to their `RawT` counterparts at all.

Furthermore, there is that issue of how I am going to do filling.

```
inl q forall t. (x : t) =
    inl w forall t. (y : t) : bool = $"(=)" x y
    w x
```

Remember this thing? Making this work would require complicating renaming significantly. It would be so hard that I am just going to decide that foralls are not allowed to shadow.

All this is a burden. All these unsolved problems I've left for later are a weight I have to carry. It is the weight of the past, bearing down upon me from the future.

9:55am. How heavy is it? About as heavy as the entire `Infer` module. So 1250 LOC heavy.

10am. I've been too arrogant.

It is one thing to make pithy plans and neatly lay it out in steps from A to Z. It is another for that plan to actual make sense.

I've been trying to move forward, but I've been ignoring the ball-and-chain around my neck and ankles. I want to signal that I am strong and that I keep moving forward, but the burden that I've ignored is wrecking me subconsciously.

10:05am. This is what will be my mission from here. I need to firmly move on from the inferencer before I reach the prepass. If I can break that dependency, the prepass will go to its expected difficulty and I will be able to handle it without much trouble.

10:10am. The same goes for the partial evaluator. Seriously, the prepass and the partial evaluator should not be this difficult! I am extremely experienced in this by now.

It should have been smooth sailing as soon as I finished the inferencer. Yet, my morale has hit rock bottom not long after.

10:15am. Though I am experienced with the prepass and partial evaluation, I've never done filling before, so it is no wonder I stumbled over a mistake.

Ok, let me take a break first and then I will start. You'd imagine the days where I program the entire day to be rough on me, but it is the last few days where I've only made tepid progress that are so.

Right now, it feels like a fire is being lit. I am feeling true motivation for the first time in a while. I'll take a little break, and then I am going to tap into it."