"1:35pm.
```fs
        | RawTLayout(r,a,b) ->
            let v = fresh_var()
            unify r s (TyLayout(v,b))
            f v a
```

Let me do this like so. But for `RawTNominal`, it will have to be an exception.

...

This is good. I think everything is set now. I found a bunch of bugs and I fixed them. And I've prepared the ground for the fill function.

Let me take a short break here.

I think everything is set so that I can finish this today. If not today, I should be able to do it tomorrow. And at that point I will have a clear path towards dealing with the rest."