"4:25pm. Yeah, it is chill.
In the end the prepass clocks in at 747 LOC.

4:30pm. Since I've come thus far, I think it is time that I start getting ready to start work on the partial evaluator. Once I have it and the codegen I will be able to move to testing of the prepass. While parsing and typechecking could be tested in isolation, this last part which is really the core of Spiral needs to be done in an integrated fashion.

Mhhhh...This next part will be time consuming to be sure. Even for the codegen although I keep saying it is not a big deal, it is still 500 LOC that I need to go through.

And the testing will take a while also. I've finished the prepass, but there will no doubt be bugs lurking in it as they will in the partial evaluator.

4:35pm.

```fs
        | ERecBlock(r,a,b) ->
            let add_term k v (env : LowerEnv) = { env with term = {|env.term with var = Map.add k v env.term.var|} }
            let a, env =
                List.mapFold (fun env (id,body) ->
                    let re = ref Unchecked.defaultof<_>
                    let body env = let x = term env body in re := x; id, x
                    body, add_term id (ERecursive re) env
                    ) env a
            let a, env = List.mapFold (fun env x -> let id,body = x env in (id,body), add_term id body env) env a
            ERecBlock(r,a,term env b)
```

Ah, wait. Yesterday I said I can just get rid of recursive blocks. I forgot about that.

```fs
        | ERecBlock(r,a,b) ->
            let add_term k v (env : LowerEnv) = { env with term = {|env.term with var = Map.add k v env.term.var|} }
            let a, env =
                List.mapFold (fun env (id,body) ->
                    let re = ref Unchecked.defaultof<_>
                    let body env = let x = term env body in re := x; id, x
                    body, add_term id (ERecursive re) env
                    ) env a
            let env = List.fold (fun env x -> let id,body = x env in add_term id body env) env a
            term env b
```

This is another bug fixed. The negative ids in bindings would definitely bring me array out of range exceptions during peval time.

4:40pm. Let me commit this as well."