"10:50am. Last night was exhausting. Not only did I go to bed late, I kicked my mind into high gear once it was time to rest.
Fiction.

Yesterday I indulged my greed a little, and once I went to bed I indulged my other desires. I was reminded that my main purpose is the self improvement loop. But it is a problem. Though I have the determination to personally go through it, there is a huge issue in that I do not have the foundation for it.

I said that the dominant strategy is not to fuse - but not to separate in the first place, but I do not feel it after all. I was bold and said that, but I do not really feel it.

I have a bunch of plans, but I have no idea how realistic any of them are. The non-separation one is the craziest. The others I doubt as well as they seem to go straight through the libidinal drives in my system.

11am. If I could see just a bit further, then my mental health would definitely improve. Having money will help a bit in the respect, but new insights into the self improvement loop are above all in importance.

In order to get them, my AI skills in the real world will need to go through a qualitative change. Once I know what is possible and what is unrealistic in reality, I will be able to put some of my regrets behind.

It is not the matter of reaching the human level of intelligence in artificial agents that is the problem here. What I do not understand is what level of control is realistic.

What I do not understand is what the knobs and the levers will be. Without that I do not know whether I will be watching a movie, or playing a RTS or FPS.

11:05am. Without that vital knowledge anything I can come up with is just indulgence. It is just fiction.

Take Reverend Insanity. I love it to death, but it is not like I am gaining much from that. Seeing Fang Yuan scheme so well might be affecting my drive to negotiate and exchange my PL for benefits, but it is not like the cultivation system in that story will have anything to do with the self improvement loop in real life.

In fact, I find the whole drive towards immortality in that story too much humanist. A person's lifespan is not some mythical quality, instead it is accumulated damage to the body. Given the kinds of healing Gu that exist in that world, dealing with that challenge should not be a big deal.

11:10am. It is just too unlikely that I am going to find any clues to the nature of reality in anything that I consume. I need computers to give me a hand.

I am really dead tired right now. But that is not too bad as I do not have the energy for stray thoughts. God knows I have too much garbage in my head, that is better off not being processed. I should be able to do some programming based on past experience of being in these situations.

Let me read a bit and then I will start. Let me get breakfast right away."